Build openstack-ligthspeed/rag-content image (#2)
* Move OpenStack scripts to openstack-lightspeed org

road-core/rag-content contains OpenStack specific scripts that
are needed if one wants to build a vector database for OpenStack
documentation.

This commit moves the scripts to the openstack-lightpseed/rag-content
repository which should from now on contain the OpenStack specific
code and road-core/rag-content should be only used as a provider
of common logic.

[1] https://github.com/road-core/rag-content

* Introduce openstack-lightspeed/rag-content image

This commit adds:

- Containerfile that can be used to build rag-content image containing
  the OpenStack vector database.

- Makefile which can be used to trigger the build of the container
  image.

* [GithubAction] Add build-and-push-rag-content workflow

This commit introduces a new workflow called build-and-push-rag-content.
This workflow builds a rag-content image which contains a vector database
containing data from OpenStack documentation. Currently, the workflow
only builds rag-content image for nova, neutron, cinder, keystone, glance
and swift projects. If the image build is successful then it is pushed
to quay.io/openstack-lightspeed/rag-content.

If the workflow got triggered by a PR creation it only builds a partial
image and does not perform a push of the image to the quay.io registry.

* Remove initialization of EXTRA_BUILD_ARGS

* Remove installation of podman

* Remove edited type from pull_request trigger

The edited type causes triggering of the job even when the PR
description gets updated.

* Remove OS_VERSION argument

The ./get_openstack_plaintext_docs.sh script does not support passing
of the OS_VERSION to the script through arguments.

* Trigger full image build on push to main

* Remove EXTRA_BUILD_ARGS

* Remove installation of jq

Jq is available in the Ubuntu runner by default.