[Github Action] Change pull_request to pull_request_target
We need to change the pull_request trigger to pull_request_target
in order to be able to access the repository secrets when PR is
proposed. By default pull_request does not have access to the
secrets and this behavior is not modifiable (due to security
reasons).

The pull_request_target trigger has access to the secrets but runs
the job against the target branch.

We checkout the code from the PR in the job which might be a
security concern but we also use a setting for this repository
which allows execution of workflows only for members of the
openstack-lightspeed org (approval required for outsiders).