Extend vector db with services present in RHOSO (#7)
* Extend vector db with services present in RHOSO

This commit extends the default set of projects we use to build
the vector database. The default value was modified to include
services available in RHOSO environment.

Signed-off-by: Lukas Piwowarski <lpiwowar@redhat.com>

* Add watcher to default projects

Let's add watcher to set of projects that are used to build the
vector db as it is gonna be soon part of RHOSO.

---------

Signed-off-by: Lukas Piwowarski <lpiwowar@redhat.com>