Use lightspeed_rag_content library (#13)
* Use lightspeed_rag_content library

road-core/rag-content introduced lightspeed_rag_content Python
package as part of this PR [1][2]. This commit updates the
generate_embeddings_openstack.py so that it uses the package.

As part of this commit:

- Containerfile is updated so that it works with the most recent
  changes in road-core/rag-content.

- The source image is changed to ghcr.io/road-core/rag-content.

- Introduces requirements.txt.

- New Python formatter is added - Black.

[1] https://github.com/road-core/rag-content/pull/25
[2] https://github.com/road-core/rag-content/pull/30

* Replace black with ruff

* Remove duplicate actions/setup-python

* Use rag-content-cpu image instead of rag-content one

The rag-content image was split into two separate images [1]:

- rag-content-cpu
- rag-content-gpu

This commit makes sure we use rag-content-cpu by default.

[1] https://github.com/road-core/rag-content/pull/36