[OSPRH-16305] Set the image tag equal to the index name (#39)
This change makes it easier to identify the index ID used within the
vector database image by tagging the image with the index ID itself.
This simplifies consumption of the vector database image, as the tag
clearly indicates which index ID should be used when consuming data from
it.

This commit also updates the tags used in our GitHub Actions pipeline.
Now, two tags will be assigned to the image:

 * Git SHA: Identifies the specific commit.

 * os-docs-${OS_VERSION}: Specifies the index ID for the data stored
   within the vector database image.