Allow setting WORKING_DIR for get docs
Using a directory in `/tmp/` may not always be a good idea, because we
are using a lot of space when building all the projects and using 2
different virtual environments for each, so it is likely that we'll run
out of space in that `tmpfs` location.

This patch allows setting `WORKING_DIR` as an environmental variable so
we can use a non `tmpfs` location if memory isn't enough in our case.