Support cleaning venv and project on get docs
This patch adds the possibility of cleaning the tox virtual env or the
whole cloned project when building the plaintext docs.

To do that we can use `CLEAN_FILES` environmental variable setting it to
`venv`, `all`, or anything else to not delete anything.

We keep current default of not cleaning things.

This is convenient if you just want to build the docs once and you
prefer to save space instead of speeding up a follow up generation.