Several fixes (#43)
* Fix get docs

On systems with more than one python version and a non 3.11 default the
`get_openstack_plaintext_docs` fail to build.

This is because the tox targets we build don't specify the python
version, so they use the system default and fail.

In this patch we force the presence of python3.11 (version can be
changed with `PYTHON_VERSION` env var) and use it for the tox targets we
create.

```
text-docs: exit 1 (2.53 seconds) /tmp/os_docs_temp/nova> python -I -m pip install -chttps://releases.openstack.org/constraints/upper/2024.2 -r /tmp/os_docs_temp/nova/doc/requirements.txt -c https://releases.openstack.org/constraints/upper/2024.2 pid=82051
```

* Allow setting WORKING_DIR for get docs

Using a directory in `/tmp/` may not always be a good idea, because we
are using a lot of space when building all the projects and using 2
different virtual environments for each, so it is likely that we'll run
out of space in that `tmpfs` location.

This patch allows setting `WORKING_DIR` as an environmental variable so
we can use a non `tmpfs` location if memory isn't enough in our case.

* Support cleaning venv and project on get docs

This patch adds the possibility of cleaning the tox virtual env or the
whole cloned project when building the plaintext docs.

To do that we can use `CLEAN_FILES` environmental variable setting it to
`venv`, `all`, or anything else to not delete anything.

We keep current default of not cleaning things.

This is convenient if you just want to build the docs once and you
prefer to save space instead of speeding up a follow up generation.