Revert "Include api-ref documentation of the OpenStack projects (#32)"
This reverts commit 6da56634259756236352021be67b5d4e4fbc131c.

Reverting the commit that introduced api-ref into the vector db image.
Unfortunately, there are a couple of issues that caused the URL links
in the vector database to be unreachable.

We are going to reintroduce this commit once it is fixed. Also, the
api-ref documentation inclusion in the image is going to be turned off
by default.

Resolves: #33