Fix consumption of rag-content-cpu image
This PR changed the base image for the rag-content-cpu image [1]. Thanks
to it we can no longer use dnf to install additional packages needed to
generate text documentation (dnf not available).

This commit fixes the issue by introducing one additional stage that is
used to generate the text formatted documentation.

Alternative solution is to use microdnf but we would need to install
more packages in the ubi-minimal stage and new issues arise like setting
the locale and new plaintext docs generation errors.

[1] https://github.com/lightspeed-core/rag-content/pull/27