Image build pass nvidia GPUs
Currently the lightspeed core GPU image is broken and doesn't work [1],
but once it gets fixed we would still be missing the right parameters
for the GPUs on the host to be used by `podman build`.

This patch passes the `--device nvidia.com/gpu=all` parameter to `podman
build` so the torch library can use it.

It also makes sure to fail if we try to use the GPU but it's not
detected inside the container, this way we make sure we notice when it
breaks.

[1]: https://issues.redhat.com/browse/LCORE-438