Fix container (#48)
* Fix container build

Our current container build mechanism is broken in multiple ways:

- Some arguments are used before being declared (eg: RHOSO_DOCS_GIT_URL)
- Some arguments are not being passed in the make target (eg:
  BUILD_UPSTREAM_DOCS)
- Some arguments are not defined in the make target (eg: OS_PROJECTS and
  DOCS_LINK_UNREACHABLE_ACTION)
- The Containerfile is assuming that `ARG` parameters become
  environmental variables and will therefore be available to script, but
  that's not the case (eg: OS_PROJECTS or OS_VERSION for
  get_openstack_plaintext_docs.sh)

This patch fixes these issues but doesn't fix the GPU build because it's
broken due to the lightspeed core base image
`ghcr.io/lightspeed-core/rag-content-gpu:latest` [1].

[1]: https://issues.redhat.com/browse/LCORE-438

* Image build pass nvidia GPUs

Currently the lightspeed core GPU image is broken and doesn't work [1],
but once it gets fixed we would still be missing the right parameters
for the GPUs on the host to be used by `podman build`.

This patch passes the `--device nvidia.com/gpu=all` parameter to `podman
build` so the torch library can use it.

It also makes sure to fail if we try to use the GPU but it's not
detected inside the container, this way we make sure we notice when it
breaks.

[1]: https://issues.redhat.com/browse/LCORE-438