Fix downstream container building
In a previous patch we fixed the upstream container building process,
but downstream was still broken.

This patch fixes the downstream build process by separating the upstream
and downstream processes of building the plain text documentation.

This approach is easier than choosing a base image and then install all
the missing dependencies.

If this is deemed to be inefficient we can always join them at a later
time.