Fix downstream container building (#49)
* Image: Speed up cloning

With this patch we speed up the cloning parts of getting the
documentation sources.

This is achieved by only cloning the branch we are interested in and
doing just a shallow cloning of the minimum depth.

This is sufficient because we don't need the git repository's history.

* Get docs scripts show executed commands

With this patch we set bash scripts int tracing mode which helps debug
errors when scripts fail on the image build process.

Scripts that now have tracing enabled are:

- get_openstack_plaintext_docs.sh
- get_rhoso_plaintext_docs.sh

* Fix downstream container building

In a previous patch we fixed the upstream container building process,
but downstream was still broken.

This patch fixes the downstream build process by separating the upstream
and downstream processes of building the plain text documentation.

This approach is easier than choosing a base image and then install all
the missing dependencies.

If this is deemed to be inefficient we can always join them at a later
time.