Make sure `OS_PROJECTS` is set to default value (#52)
* Make sure OS_PROJECTS is set to default value

The OS_PROJECTS variable is accessed when it is not set. If it is not
set through ENV variable running the get_openstack_plaintext.sh script
outside of the container can fail:

./scripts/get_openstack_plaintext_docs.sh: line 47: OS_PROJECTS: unbound
variable

This commit makes sure that OS_PROJECTS is set even when it is not
passed through the ENV variable. This enables running of the script
outside of the Container.

* Introduce sane behavior when OS_PROJECTS is empty

Make sure that OS_PROJECTS is not set to the default value specified
within the script when it is set to an empty string.