[OSPRH-19820] Do not use commit tag for periodic builds
The openstack-k8s-operators/openstack-operator references the
openstack-lightspeed/rag-content image. During the build process of the
openstack-operator, the os-docs-2024.2 tag gets transferred to a sha256
value (the openstack-operator build references the image through that
sha256 value).

Until now, this repository was overwriting both the image with the
commit-tag and the os-docs-2024.2 tag in the periodic build, which removed
the image with the sha256 value that was referenced in the earlier
openstack-operator build.

This commit fixes this issue by using nightly-YYYYMMDD tags for the
periodic builds to ensure that existing tags do not get overwritten
and the referenced images with sha256 values do not disappear.