Build OCP RAG DB
In this patch we include the mechanism to build RAG DBs for OCP.

It retrieves the docs from the upstream doc repository, then for each of
the selected OCP versions (or all) it removes unwanted files and
directories, and then it builds a RAG DB for each version together with
common docs for all OCP versions.

Directories and files to exclude are in the
`scripts/ocp_exclude_docs.conf` file, which allows lines to start with
`#` for comments.

NOTE: The OLS documentation seems to have some variables that have
      not been properly replaced when doing the asciidoc conversion.
      For example `{product-title}`, `{product-version}` and
      `{FeatureName}`.  This patch does not fix these issues.