Fix GPU container flavor
Even though the GPU base image from lightspeed-core has been fixed our
image still fails to build using the GPU.

Failure happens on the embedding calculation phase, and it happens
because we are copying `/rag-content` from other phases, and this
directory contains torch libraries.  In those previous phases we were
using the CPU image because we didn't need the GPU, so we end up
replacing the GPU libraries with the CPU ones, and that's why it fails.

In this patch we fix this by ensuring the right image is used in all the
image building phases.